@@ -18,7 +18,8 @@ Gem::Specification.new do |s|
 
   s.required_ruby_version = '>= 2.2.2'
 
-  s.add_dependency "rails", "~> 5.0"
+  s.add_dependency "rails", "~> 5.2"
+  s.add_dependency 'jquery-rails', '~> 4.3'
   s.add_dependency "kaminari"
   s.add_dependency "nokogiri"
   s.add_dependency "fastimage"
