@@ -1,4 +1,4 @@
-class AddPageToSearches < ActiveRecord::Migration
+class AddPageToSearches < ActiveRecord::Migration[4.2]
   def change
     add_column :searches, :page, :string
   end
