@@ -1,4 +1,4 @@
-class CreateSessions < ActiveRecord::Migration
+class CreateSessions < ActiveRecord::Migration[5.0]
   def change
     create_table :sessions do |t|
       t.string :session_uuid
